requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
wordcloud==1.6.0
